# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Ascension Quest API
  version: 0.1.0
  description: "Ascension Quest: Legends of Aethoria API"
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nutrition/targets:
    get:
      operationId: getNutritionTargets
      tags: [nutrition]
      summary: Get nutrition targets
      responses:
        "200":
          description: Nutrition targets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionTargets"
    patch:
      operationId: updateNutritionTargets
      tags: [nutrition]
      summary: Update nutrition targets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionTargets"
      responses:
        "200":
          description: Updated targets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionTargets"
...
  /nutrition/logs:
    get:
      operationId: getNutritionLogs
      tags: [nutrition]
      summary: Get nutrition logs
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Nutrition logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NutritionLog"
    post:
      operationId: createNutritionLog
      tags: [nutrition]
      summary: Log a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionLogInput"
      responses:
        "201":
          description: Created log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLog"
  /nutrition/logs/{id}:
    patch:
      operationId: updateNutritionLog
      tags: [nutrition]
      summary: Update a nutrition log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionLogUpdate"
      responses:
        "200":
          description: Updated log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLog"
    delete:
      operationId: deleteNutritionLog
      tags: [nutrition]
      summary: Delete a nutrition log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /nutrition/meals/saved:
    get:
      operationId: getSavedMeals
      tags: [nutrition]
      summary: List saved meals
      responses:
        "200":
          description: Saved meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedMeal"
    post:
      operationId: createSavedMeal
      tags: [nutrition]
      summary: Save a meal template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedMealInput"
      responses:
        "201":
          description: Saved meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedMeal"
  /nutrition/meals/saved/{id}:
    delete:
      operationId: deleteSavedMeal
      tags: [nutrition]
      summary: Delete a saved meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /nutrition/weight:
    get:
      operationId: getWeightEntries
      tags: [nutrition]
      summary: Get weight log entries
      responses:
        "200":
          description: Weight entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightEntry"
    post:
      operationId: createWeightEntry
      tags: [nutrition]
      summary: Log body weight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightEntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightEntry"

  # ─── TRAINING ─────────────────────────────────────────────────────────────────
...
  /nutrition/food-search:
    get:
      operationId: searchFood
      tags: [nutrition]
      summary: Search food database (Open Food Facts)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Food search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodSearchResult"
...
components:
  schemas:
...
    NutritionTargets:
      type: object
      required: [calories, protein, carbs, fat]
      properties:
        calories:
          type: integer
        protein:
          type: integer
        carbs:
          type: integer
        fat:
          type: integer
        sex:
          type: string
          enum: [male, female]
          nullable: true
        ageYears:
          type: integer
          nullable: true
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active, very_active]
          nullable: true
        weightGoal:
          type: string
          enum: [lose, maintain, gain]
          nullable: true
        autoCalc:
          type: boolean
...
    NutritionLog:
      type: object
      required: [id, date, mealName, calories, protein, carbs, fat, mealType, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        mealName:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack, pre_workout, post_workout]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    NutritionLogInput:
      type: object
      required: [mealName, calories, protein, carbs, fat, mealType]
      properties:
        date:
          type: ["string", "null"]
        mealName:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack, pre_workout, post_workout]
        notes:
          type: ["string", "null"]
    NutritionLogUpdate:
      type: object
      properties:
        mealName:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack, pre_workout, post_workout]
        notes:
          type: ["string", "null"]
    SavedMeal:
      type: object
      required: [id, name, calories, protein, carbs, fat]
      properties:
        id:
          type: integer
        name:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        mealType:
          type: ["string", "null"]
    SavedMealInput:
      type: object
      required: [name, calories, protein, carbs, fat]
      properties:
        name:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        mealType:
          type: ["string", "null"]
...
    WeightEntry:
      type: object
      required: [id, weight, date, unit]
      properties:
        id:
          type: integer
        weight:
          type: number
        date:
          type: string
        unit:
          type: string
          enum: [lbs, kg]
        notes:
          type: ["string", "null"]
    WeightEntryInput:
      type: object
      required: [weight, date, unit]
      properties:
        weight:
          type: number
        date:
          type: string
        unit:
          type: string
          enum: [lbs, kg]
        notes:
          type: ["string", "null"]

    # ─── TRAINING ──────────────────────────────────────────────────────────────
...
    FoodSearchResult:
      type: object
      required: [id, name, calories100g, protein100g, carbs100g, fat100g]
      properties:
        id:
          type: string
        name:
          type: string
        calories100g:
          type: number
        protein100g:
          type: number
        carbs100g:
          type: number
        fat100g:
          type: number
        fiber100g:
          type: ["number", "null"]
        sugar100g:
          type: ["number", "null"]
        sodiumMg100g:
          type: ["number", "null"]
        servingSize:
          type: ["string", "null"]
        servingGrams:
          type: ["number", "null"]
        category:
          type: string
        source:
          type: string
          enum: [guild_compendium, open_food_facts]
...